completely separated all components into their own files and added react-router-dom routing
