+++ package.json
@@ -10,6 +10,7 @@
     "@testing-library/user-event": "^12.8.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.1",
     "react-scripts": "4.0.3",
     "rellax": "^1.12.1",
     "web-vitals": "^1.1.2"
