+++ src/components/CollageComponent.js
