+++ src/components/Footer.js
@@ -1,0 +1,9 @@
+const Footer = () => (
+  <footer className="absolute w-full m-auto z-10 bg-purple-even-darker">
+    <div className="relative text-center object-center content-center items-center m-auto my-10">
+      <span className="text-sm font-normal text-indigo-200">Made with ❤️ by Fred Old – Code available on <a className="underline" target="_blank" rel="noreferrer" href="https://github.com/Vurak/vurak.github.io/tree/react-dev-branch">GitHub</a></span>
+    </div>
+  </footer>
+)
+
+export default Footer
