+++ src/routes/index.js
@@ -1,0 +1,26 @@
+import text from "../text"
+const routes = [
+  {
+    path: '/game',
+    textObject: text.projects.game
+  },
+  {
+    path: '/lmgtfy',
+    textObject: text.projects.lmgtfy
+  },
+  {
+    path: '/blender',
+    textObject: text.projects.blender
+  },
+  {
+    path: '/sigmo',
+    textObject: text.projects.sigmo
+  },
+  {
+    path: '/portfolio',
+    textObject: text.projects.portfolio
+  },
+
+]
+
+export default routes
